// src/common/enum/user.enum.ts
export enum RoleEnum {
    user = "user",
    admin = "admin",
}


export enum statusAccountEnum {
    active = "active",
    inactive = "inactive",
}


export enum typeTransactionEnum {
    deposit = "deposit",
    withdrawal = "withdrawal",
    transfer = "transfer",
}

export enum statusTransactionEnum {
    pending = "pending",
    success = "success",
    failed = "failed",
}